Some order in test base classes
